+++ app/controllers/account/orders_controller.rb
@@ -4,19 +4,5 @@ class Account::OrdersController < ApplicationController
   def index
     @orders = current_user.orders.order("id DESC")
   end
-  def pay_with_alipay
-    @order = Order.find_by_token(params[:id])
-    @order.set_payment_with!("alipay")
-    @order.pay!
-
-    redirect_to order_path(@order.token), notice: "使用支付宝成功完成付款"
-  end
-
-  def pay_with_wechat
-    @order = Order.find_by_token(params[:id])
-    @order.set_payment_with!("wechat")
-    @order.pay!
-
-    redirect_to order_path(@order.token), notice: "使用微信成功完成付款"
-  end
+  
 end
