+++ config/routes.rb
@@ -18,13 +18,14 @@ Rails.application.routes.draw do
      end
   end
    resources :cart_items
-   resources :orders
-  namespace :account do
    resources :orders do
       member do
        post :pay_with_alipay
        post :pay_with_wechat
       end
     end
+  namespace :account do
+    resources :orders
+
   end
 end
