+++ test/mailers/order_mailer_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class OrderMailerTest < ActionMailer::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
